Develop

+++ src/main/java/com/ebarapp/ebar/service/VotingService.java
@@ -33,7 +33,7 @@ public class VotingService {
     }
 
     @Transactional(readOnly = true)
-    public List<Voting> getVotingsByBarId(Long barId) {
-        return votingRepository.getVotingsByBarId(barId.intValue());
+    public List<Voting> getVotingsByBarId(Integer barId) {
+        return votingRepository.getVotingsByBarId(barId);
     }
 }
